LPS-186688 Rename derived classes to match base class
