+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/display/context/FolderActionDisplayContext.java
@@ -365,7 +365,7 @@ public class FolderActionDisplayContext {
 		return PortletURLBuilder.createRenderURL(
 			_dlRequestHelper.getLiferayPortletResponse()
 		).setMVCRenderCommandName(
-			"/document_library/copy_folder"
+			"/document_library/folder_copy_entry"
 		).setRedirect(
 			_dlRequestHelper.getCurrentURL()
 		).setParameter(
