+++ modules/apps/document-library/document-library-web/src/main/resources/META-INF/resources/document_library/file_entry_copy_entry.jsp
